chore(constants): move constants from branding file
