+++ apps/native/src/components/widget/settings/ai-models-tab.tsx
@@ -9,7 +9,7 @@ import {
 } from "@/components/ui/select";
 import { ModelCombobox } from "@/components/widget/controls/model-combobox";
 import { getProviderConfigInvalidReason, isCliProvider } from "@/lib/ai-provider-validation";
-import { DEFAULT_MAX_ITERATIONS } from "@/../shared/constants";
+import { DEFAULT_MAX_ITERATIONS } from "@/lib/constants";
 import { darwinAPI } from "@/tauri-api";
 import type { CliToolsState } from "@/types/shared";
 import type { AnyFieldApi, ReactFormExtendedApi } from "@tanstack/react-form";
