+++ apps/native/src/lib/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_MAX_ITERATIONS = 25;
+export const EVOLVE_EVENT_CHANNEL = "darwin:evolve:event";
